- implemented find_trade, find_order methods in connector - added base implementation of EMA strategy - refactored oanda connector - fixed small bugs in oanda connector

+++ Platform/tbp/oanda/include/oanda/connector.h
@@ -23,7 +23,7 @@ namespace tbp
 			~authentication();
 		};
 
-		data_t create_market_order_params(const std::wstring& instrument_id, long amount);
+		data_t create_market_order_params(const std::wstring& instrument_id, double amount);
 
 		/////////////////////////////////////////////////////////////////////////
 		// connector
